CASSANDRA-7438

+++ src/java/org/apache/cassandra/cache/InstrumentingCache.java
@@ -94,7 +94,7 @@ public class InstrumentingCache<K, V>
         capacitySetManually = true;
     }
 
-    public int size()
+    public long size()
     {
         return map.size();
     }
